+++ src/types/index.ts
@@ -1,6 +1,11 @@
 export type { StyleSpecification } from '@maplibre/maplibre-gl-style-spec';
 export type { MaplibreFilter, MaplibreLayer, MaplibreLayerDefinition } from './maplibre.js';
 export type { TileJSONSpecification, TileJSONSpecificationRaster, TileJSONSpecificationVector } from './tilejson.js';
-export { isTileJSONSpecification, isRasterTileJSONSpecification } from './tilejson.js';
+export {
+	assertTileJSONSpecification,
+	assertRasterTileJSONSpecification,
+	isTileJSONSpecification,
+	isRasterTileJSONSpecification,
+} from './tilejson.js';
 export type { VectorLayer } from './vector_layer.js';
 export { isVectorLayer, isVectorLayers } from './vector_layer.js';